pyproject.toml
[project]
name = "azure-functions-test"
dynamic = ["version"]
description = "Unit test Azure Functions without the runtime"
readme = "README.md"
license = { text = "MIT" }
authors = [{ name = "Sudarshan", email = "your@email.com" }]
requires-python = ">=3.11"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Framework :: Pytest",
    "Topic :: Software Development :: Testing",
    "Typing :: Typed",
]
keywords = ["azure", "functions", "testing", "pytest", "mock", "serverless"]
dependencies = [
    "azure-functions>=1.17.0",
    "pydantic>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.0",
    "pyright>=1.1.350",
    "ruff>=0.3",
    "bandit>=1.7",
    "pip-audit>=2.7",
]
docs = [
    "mkdocs>=1.5",
    "mkdocs-material>=9.5",
    "mkdocstrings[python]>=0.24",
]

# [project.entry-points.pytest11]
# azure_functions_test = "azure_functions_test.fixtures.pytest_plugin"

[project.urls]
Homepage = "https://github.com/yourusername/azure-functions-test"
Documentation = "https://yourusername.github.io/azure-functions-test"
Repository = "https://github.com/yourusername/azure-functions-test"
Changelog = "https://github.com/yourusername/azure-functions-test/blob/main/CHANGELOG.md"

[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[tool.hatch.version]
source = "vcs"
fallback-version = "0.0.0"

[tool.hatch.build.hooks.vcs]
version-file = "src/azure_functions_test/_version.py"

[tool.hatch.build.targets.wheel]
packages = ["src/azure_functions_test"]

# ==============================================================================
# Ruff Configuration (Linting + Formatting)
# ==============================================================================
[tool.ruff]
target-version = "py311"
line-length = 88
src = ["src", "tests", "examples"]
exclude = [
    "_version.py",  # Auto-generated by hatch-vcs
]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors (includes E501 line-too-long)
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "UP",   # pyupgrade
    "ARG",  # flake8-unused-arguments
    "SIM",  # flake8-simplify
    "PTH",  # flake8-use-pathlib
    "RET",  # flake8-return
    "TRY",  # tryceratops
]
ignore = [
    "TRY003",  # Avoid specifying long messages outside exception class
]

# Enable line length errors (E501) to show in editor
# Note: Ruff will auto-format on save, but will show error before save
[tool.ruff.lint.pycodestyle]
max-line-length = 88

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = [
    "ARG",  # Allow unused arguments in tests
    "S101", # Allow assert in tests
]

# ==============================================================================
# Pyright Configuration (Type Checking)
# ==============================================================================
[tool.pyright]
pythonVersion = "3.11"
typeCheckingMode = "strict"
include = ["src", "tests", "examples"]
exclude = ["**/__pycache__", "build", "dist", ".venv"]

# Strict mode settings
reportMissingImports = true
reportMissingTypeStubs = true
reportUnusedImport = true
reportUnusedClass = true
reportUnusedFunction = true
reportUnusedVariable = true
reportDuplicateImport = true
reportPrivateUsage = true
reportConstantRedefinition = true
reportIncompatibleMethodOverride = true
reportIncompatibleVariableOverride = true
reportInconsistentConstructor = true
reportUnnecessaryIsInstance = true
reportUnnecessaryCast = true
reportUnnecessaryComparison = true
reportUnnecessaryContains = true
reportUnusedCallResult = false
reportUnknownMemberType = true

# ==============================================================================
# Pytest Configuration
# ==============================================================================
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-ra",                    # Show summary of all test outcomes
    "-q",                     # Quiet mode
    "--strict-markers",       # Strict marker usage
    "--strict-config",        # Strict config
    "--cov=azure_functions_test",  # Coverage for our package
    "--cov-report=term-missing",   # Show missing lines in terminal
    "--cov-report=html",           # Generate HTML coverage report
    "--cov-report=xml",            # Generate XML for CI
]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "slow: Slow tests",
]

# ==============================================================================
# Coverage Configuration
# ==============================================================================
[tool.coverage.run]
source = ["azure_functions_test"]
branch = true
omit = [
    "*/tests/*",
    "*/_version.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "@abstractmethod",
    "@overload",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]
precision = 2
show_missing = true
skip_covered = false

[tool.coverage.html]
directory = "htmlcov"

# ==============================================================================
# Bandit Configuration (Security)
# ==============================================================================
[tool.bandit]
exclude_dirs = ["tests", ".venv"]
skips = ["B101"]  # Allow assert in non-test code for type narrowing
